backend/requirements.txt
flask
flask-cors
torch
torchvision
pillow
tqdm
timm
opencv-python
torch
torchvision